Clean custom code by adding gorilla mux router
