+++ app/go.mod
@@ -6,6 +6,7 @@ require github.com/lib/pq v1.10.7
 
 require (
 	github.com/afaguilarr/go-example-webserver/proto v0.0.0-20221026005202-d2489f2268d4
+	github.com/gorilla/mux v1.8.0
 	github.com/joho/godotenv v1.4.0
 	github.com/pkg/errors v0.9.1
 	github.com/pressly/goose/v3 v3.7.0
